l10n_br_cnpj_search: merge webservice classes
resolve pre-commit:
l10n_br_cnpj_search/models/serpro_webservice.py:19:
[R8180(consider-merging-classes-inherited), SerproWebservice]
Consider merging classes inherited to "l10n_br_cnpj_search.webservice.abstract" from l10n_br_cnpj_search/models/receitaws_webservice.py:11:4.
