+++ l10n_br_cnpj_search/models/__init__.py
@@ -3,7 +3,5 @@
 from . import l10n_br_base_party_mixin
 from . import res_config_settings
 from . import cnpj_webservice
-from . import receitaws_webservice
-from . import serpro_webservice
 from . import res_partner
 from . import res_company
